# backend/openapi/v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "XUGOU v2 API",
    "version": "2.0.0",
    "description": "Single-Worker modular-monolith API contract."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v2/notifications": {
      "get": {
        "operationId": "getNotificationConfigV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NotificationConfig"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/notifications/channels": {
      "get": {
        "operationId": "listNotificationChannelsV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "maxItems": 100,
                      "items": {
                        "$ref": "#/components/schemas/NotificationChannel"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createNotificationChannelV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CreatedId"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationChannelCommand"
              }
            }
          }
        }
      }
    },
    "/api/v2/notifications/channels/{id}": {
      "get": {
        "operationId": "getNotificationChannelV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NotificationChannel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ]
      },
      "patch": {
        "operationId": "updateNotificationChannelV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CreatedId"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationChannelMutation"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteNotificationChannelV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ]
      }
    },
    "/api/v2/notifications/channels/{id}/test": {
      "post": {
        "operationId": "testNotificationChannelV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "delivered"
                      ],
                      "properties": {
                        "delivered": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ]
      }
    },
    "/api/v2/notifications/templates": {
      "get": {
        "operationId": "listNotificationTemplatesV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "maxItems": 100,
                      "items": {
                        "$ref": "#/components/schemas/NotificationTemplate"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createNotificationTemplateV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CreatedId"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationTemplateCommand"
              }
            }
          }
        }
      }
    },
...
    "/api/v2/notifications/settings": {
      "put": {
        "operationId": "saveNotificationSettingV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CreatedId"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettingCommand"
              }
            }
          }
        }
      }
    },
...
    "/api/v2/notifications/history": {
      "get": {
        "operationId": "listNotificationHistoryV2",
        "tags": [
          "notifications"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationHistoryPage"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "monitor",
                "agent"
              ]
            }
          },
          {
            "name": "target_id",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "success",
                "failed"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "xugou_session"
      },
      "releaseToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque release credential"
      },
      "agentBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "CreatedId": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      },
...
      "NotificationConfig": {
        "type": "object",
        "required": [
          "channels",
          "templates",
          "channels_has_more",
          "templates_has_more",
          "settings"
        ],
        "properties": {
          "channels": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/NotificationChannel"
            }
          },
          "templates": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/NotificationTemplate"
            }
          },
          "channels_has_more": {
            "type": "boolean"
          },
          "templates_has_more": {
            "type": "boolean"
          },
          "settings": {
            "type": "object",
            "required": [
              "monitors",
              "agents",
              "specificMonitors",
              "specificAgents"
            ],
            "properties": {
              "monitors": {
                "$ref": "#/components/schemas/NotificationMonitorSetting"
              },
              "agents": {
                "$ref": "#/components/schemas/NotificationAgentSetting"
              },
              "specificMonitors": {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/NotificationMonitorSetting"
                }
              },
              "specificAgents": {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/NotificationAgentSetting"
                }
              }
            }
          }
        }
      },
...
    }
  }
}